+++ R/rfcm--presets.R
@@ -4,7 +4,7 @@
 default_rfms_map <- function(seed = 840) {
 
   withr::with_seed(seed, {
-    new_rfms_map(default_management_system = jsendra()) |>
+    new_rfms_map(default_rfms = jsendra()) |>
       allocate_surface(clearfield(), target_fraction = 0.8, ditches = 1:9) |>
       allocate_surface(clearfield(), target_fraction = 0.2, ditches = 10:26)
     })