+++ R/rfcm--s3-class.R
@@ -5,9 +5,9 @@
 #' associated with a specific rice field management system, allowing for
 #' heterogeneous management across the simulation domain.
 #'
-#' @param default_management_system `[`[erahumed_management_system][management_system]`]` \cr
+#' @param default_rfms `[`[erahumed_rfms][rfms]`]` \cr
 #'   A default management system assigned to all clusters initially. Typically
-#'   created with [new_management_system()] or a helper like [jsendra()].
+#'   created with [new_rfms()] or a helper like [jsendra()].
 #' @param seed `[numeric(1)]` \cr
 #'  Seed for random number generation in the assignation of clusters to
 #'  management systems.
@@ -27,20 +27,20 @@
 #'
 #' Cluster assignments can be modified using [allocate_surface()].
 #'
-#' @seealso [new_management_system()], [allocate_surface()]
+#' @seealso [new_rfms()], [allocate_surface()]
 #'
 #' @name rfms_map
 #'
 #' @export
-new_rfms_map <- function(default_management_system = new_management_system())
+new_rfms_map <- function(default_rfms = new_rfms())
 {
   map_df <- data.frame(
     cluster_id = info_clusters()$element_id,
     rfms_id = 1,
-    rfms_name = default_management_system[["display_name"]]
+    rfms_name = default_rfms[["display_name"]]
     )
 
-  res <- list(map_df = map_df, rfms_list = list(default_management_system))
+  res <- list(map_df = map_df, rfms_list = list(default_rfms))
 
   class(res) <- "erahumed_rfms_map"