+++ tests/testthat/test-rfms--plot.R
@@ -3,5 +3,5 @@ test_that("Plot succeeds with Clearfield", {
 })
 
 test_that("Plot succeeds with Empty management system", {
-  expect_no_error(plot_rfms(new_management_system()))
+  expect_no_error(plot_rfms(new_rfms()))
 })